third_party/blink/web_tests/wpt_internal/wake-lock/wakelockpermissiondescriptor.https.html

<!DOCTYPE html>
<link rel="help" href="https://w3c.github.io/wake-lock/#the-wakelockpermissiondescriptor-dictionary">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>
// TODO(https://crbug.com/1015327): Move this to WPT once we can change
// permissions via testdriver.js.

promise_test(async t => {
  await test_driver.set_permission({name: 'wake-lock', type: 'screen'}, 'denied', false);

  return navigator.permissions.query({name:'wake-lock', type: 'screen'}).then(status => {
    assert_class_string(status, "PermissionStatus");
    assert_equals(status.state, "denied");
  });
}, "WakeLockPermissionDescriptor with type=screen works");

promise_test(async t => {
  await test_driver.set_permission({name: 'wake-lock', type: 'system'}, 'denied', false);

  return navigator.permissions.query({ name: 'wake-lock', type: 'system' }).then(status => {
    assert_class_string(status, "PermissionStatus");
    assert_equals(status.state, "denied");
  });
}, "WakeLockPermissionDescriptor with type=system works");
</script>
